+++ scripts/run_tile_registration.py
@@ -4,7 +4,7 @@ import configparser
 from prefect import Flow, Parameter, unmapped
 from prefect.executors import LocalDaskExecutor
 
-from sbem.tile_stitching.sofima_utils import (
+from sbem.tile_stitching.sofima_tasks import (
     build_integration_config,
     load_sections,
     run_sofima,
